Bugfix for queries where simple 'is a' clauses that return no matches were incorrectly being ignored.
This is demonstrated by the two new rules 'grandparent of patient' and 'grandparent of patient - long'.  Both of these should return the same result (no rows).  Prior to this fix the 'grandparent of patient - long' rule was incorrectly ignoring the last clause because no rows were matched and this then incoorectly returned results as if this clause was not present.