Removed old conditional logic that would not add an affectedConcept to a source if the instance in question was already registered as that concept.
For example the sentence 'the person Fred is a patient.' would previously only register 'patient' to affectedConcepts (on CeSource) because Fred was known to be a person but not a patient.  This meant that any triggers that were waiting to be notified about changes to person instances would not be triggered.

The removal of the conditional logic now means that both 'person' and 'patient' are registered as affected concepts, removing this issue.